VS2017 typecast fixes

+++ src/output/direct3d/direct3d.cpp
@@ -281,8 +281,8 @@ bool CDirect3D::UnlockTexture(void)
             if(!(index & 1)) {
                 y += changedLines[index];
             } else {
-                rect.top = y;
-                rect.bottom = y + changedLines[index];
+                rect.top = (LONG)y;
+                rect.bottom = (LONG)(y + changedLines[index]);
                 lpTexture->AddDirtyRect(&rect);
                 y += changedLines[index];
             }
@@ -1053,17 +1053,17 @@ HRESULT CDirect3D::Resize3DEnvironment(Bitu window_width, Bitu window_height, Bi
 	(void)fullscreen; // FIXME: This should be stored and used!
 
     // set the presentation parameters
-	d3dpp.BackBufferWidth = window_width;
-	d3dpp.BackBufferHeight = window_height;
+	d3dpp.BackBufferWidth = (UINT)window_width;
+	d3dpp.BackBufferHeight = (UINT)window_height;
 
-    dwScaledWidth = width;
-    dwScaledHeight = height;
+    dwScaledWidth = (DWORD)width;
+    dwScaledHeight = (DWORD)height;
 
-    dwX = x;
-    dwY = y;
+    dwX = (DWORD)x;
+    dwY = (DWORD)y;
 
-    dwWidth = rwidth;
-    dwHeight = rheight;
+    dwWidth = (DWORD)rwidth;
+    dwHeight = (DWORD)rheight;
 
 #if LOG_D3D
     LOG_MSG("D3D:Resolution set to %dx%d%s", d3dpp.BackBufferWidth, d3dpp.BackBufferHeight, fullscreen ? ", fullscreen" : "");
